Use options from request for settings
