+++ lib/connectors/memory.js
@@ -323,7 +323,7 @@ Memory.prototype.upsertWithWhere = function(model, where, data, options, callbac
   }
 };
 
-Memory.prototype.findOrCreate = function(model, filter, data, callback) {
+Memory.prototype.findOrCreate = function(model, filter, data, options, callback) {
   var self = this;
   var nodes = self._findAllSkippingIncludes(model, filter);
   var found = nodes[0];
@@ -345,7 +345,7 @@ Memory.prototype.findOrCreate = function(model, filter, data, callback) {
     });
   }
 
-  self._models[model].model.include(nodes[0], filter.include, {}, function(err, nodes) {
+  self._models[model].model.include(nodes[0], filter.include, options, function(err, nodes) {
     process.nextTick(function() {
       if (err) return callback(err);
       callback(null, nodes[0], false);
