+++ lib/include.js
@@ -205,8 +205,9 @@ Inclusion.include = function(objects, include, options, cb) {
    */
   function findWithForeignKeysByPage(model, filter, fkName, pageSize, options, cb) {
     try {
-      model._sanitizeQuery(filter.where, {prohibitProtectedPropertiesInQuery: true});
-      model._coerce(filter.where);
+      const opts = Object.assign({prohibitProtectedPropertiesInQuery: true}, options);
+      model._sanitizeQuery(filter.where, opts);
+      model._coerce(filter.where, options);
     } catch (e) {
       return cb(e);
     }
