+++ lib/model.js
@@ -20,6 +20,8 @@ var g = require('strong-globalize')();
 var util = require('util');
 var jutil = require('./jutil');
 var List = require('./list');
+var DataAccessUtils = require('./model-utils');
+var Observer = require('./observer');
 var Hookable = require('./hooks');
 var validations = require('./validations');
 var _extend = util._extend;
@@ -549,27 +551,29 @@ ModelBaseClass.prototype.toObject = function(onlySchema, removeHidden, removePro
 };
 
 /**
+ * Convert an array of strings into an object as the map
+ * @param {string[]} arr An array of strings
+ */
+function asObjectMap(arr) {
+  var obj = {};
+  if (Array.isArray(arr)) {
+    for (var i = 0; i < arr.length; i++) {
+      obj[arr[i]] = true;
+    }
+    return obj;
+  }
+  return arr || obj;
+}
+/**
  * Checks if property is protected.
  * @param {String} propertyName Property name
  * @returns  {Boolean} true or false if protected or not.
  */
 ModelBaseClass.isProtectedProperty = function(propertyName) {
-  var Model = this;
-  var settings = Model.definition && Model.definition.settings;
-  var protectedProperties = settings && (settings.protectedProperties || settings.protected);
-  if (Array.isArray(protectedProperties)) {
-    // Cache the protected properties as an object for quick lookup
-    settings.protectedProperties = {};
-    for (var i = 0; i < protectedProperties.length; i++) {
-      settings.protectedProperties[protectedProperties[i]] = true;
-    }
-    protectedProperties = settings.protectedProperties;
-  }
-  if (protectedProperties) {
-    return protectedProperties[propertyName];
-  } else {
-    return false;
-  }
+  var settings = (this.definition && this.definition.settings) || {};
+  var protectedProperties = settings.protectedProperties || settings.protected;
+  settings.protectedProperties = asObjectMap(protectedProperties);
+  return settings.protectedProperties[propertyName];
 };
 
 /**
@@ -578,22 +582,10 @@ ModelBaseClass.isProtectedProperty = function(propertyName) {
  * @returns {Boolean} true or false if hidden or not.
  */
 ModelBaseClass.isHiddenProperty = function(propertyName) {
-  var Model = this;
-  var settings = Model.definition && Model.definition.settings;
-  var hiddenProperties = settings && (settings.hiddenProperties || settings.hidden);
-  if (Array.isArray(hiddenProperties)) {
-    // Cache the hidden properties as an object for quick lookup
-    settings.hiddenProperties = {};
-    for (var i = 0; i < hiddenProperties.length; i++) {
-      settings.hiddenProperties[hiddenProperties[i]] = true;
-    }
-    hiddenProperties = settings.hiddenProperties;
-  }
-  if (hiddenProperties) {
-    return hiddenProperties[propertyName];
-  } else {
-    return false;
-  }
+  var settings = (this.definition && this.definition.settings) || {};
+  var hiddenProperties = settings.hiddenProperties || settings.hidden;
+  settings.hiddenProperties = asObjectMap(hiddenProperties);
+  return settings.hiddenProperties[propertyName];
 };
 
 ModelBaseClass.prototype.toJSON = function() {
@@ -840,13 +832,15 @@ ModelBaseClass.getMergePolicy = function(options) {
  */
 
 ModelBaseClass.getUpdateOnlyProperties = function() {
-  var Model = this;
   const props = this.definition.properties;
   return Object.keys(props).filter(key => props[key].updateOnly);
 };
 
+// Mix in utils
+jutil.mixin(ModelBaseClass, DataAccessUtils);
+
 // Mixin observer
-jutil.mixin(ModelBaseClass, require('./observer'));
+jutil.mixin(ModelBaseClass, Observer);
 
 jutil.mixin(ModelBaseClass, Hookable);
 jutil.mixin(ModelBaseClass, validations.Validatable);
